Fixed data and timestamps to retrieve the data correctly

+++ SmartHomeLightingSystem/src/main/java/com/smarthomelightingsystem/model/LDRData.java
@@ -7,7 +7,7 @@ public class LDRData {
 	/**
 	 * List of data values retrieved from LDR
 	 */
-    List<Integer> data = new ArrayList<>();
+    List<Float> data = new ArrayList<>();
     /**
      * List of timestamps
      */
@@ -18,11 +18,11 @@ public class LDRData {
     public LDRData() {
     }
 
-    public List<Integer> getData() {
+    public List<Float> getData() {
         return data;
     }
 
-    public void setData(List<Integer> data) {
+    public void setData(List<Float> data) {
         this.data = data;
     }
